Refactor endpoint trace emit helpers
